WIP - start fix
Beggining of an attempt to report a compile time SDE
when a typeCalc DPath expression refers to an invalid type.

I am abandoning this attempt because it appears to be more
effort that originally thought. The core issue that that
we would need the DPath expression compiler to have access
to the typeCalcMap. However, this introduces a circular dependency
which is not easy to solve.